Revert "Fixed bug when audio players didn't close in resource recordable fragment (#267)"
This reverts commit 97489db00f7a2353da2a33d0be79a53b6dbcbd56.
