+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/viewmodel/RecordResourceViewModel.kt
@@ -153,12 +153,10 @@ class RecordResourceViewModel : ViewModel() {
     }
 
     fun nextChunk() {
-        recordableViewModel.closePlayers()
         stepToChunk(StepDirection.FORWARD)
     }
 
     fun previousChunk() {
-        recordableViewModel.closePlayers()
         stepToChunk(StepDirection.BACKWARD)
     }
