+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/viewmodel/RecordScriptureViewModel.kt
@@ -68,12 +68,10 @@ class RecordScriptureViewModel : ViewModel() {
     }
 
     fun nextChunk() {
-        recordableViewModel.closePlayers()
         stepToChunk(StepDirection.FORWARD)
     }
 
     fun previousChunk() {
-        recordableViewModel.closePlayers()
         stepToChunk(StepDirection.BACKWARD)
     }
